+++ src/openeo_aggregator/about.py
@@ -2,7 +2,7 @@ import logging
 import sys
 from typing import Optional
 
-__version__ = "0.12.3a1"
+__version__ = "0.13.0a1"
 
 
 def log_version_info(logger: Optional[logging.Logger] = None):