+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is roughly based on [Keep a Changelog](https://keepachangelog.com/en/
 - Add (optional) config for collection id whitelisting.
   Keep union of all "upstream" collections as default.
   ([#129](https://github.com/Open-EO/openeo-aggregator/issues/129))
+- Disable `auto_validation` feature of latest `openeo` python client library release ([#130](https://github.com/Open-EO/openeo-aggregator/issues/130))
 
 
 ## [0.11.x]